add start stop

+++ src/rviz_plugin/include/waypoint_nav_frame.h
@@ -145,7 +145,7 @@ private Q_SLOTS:
   void setMission(int index);
   void topicChanged();
   void poseChanged(double val);
-  void speedChanged(int val);
+  void speedChanged(double val);
   void saveButtonClicked();
   void loadButtonClicked();
   void makeWPButtonClicked();
